add slider with seek to + show timestamp

+++ MAC_Slider_src/MAC_Slider_C_Sharp_src/MACTrackBarLib/MACTrackBar.cs
@@ -56,7 +56,7 @@ namespace XComponent.SliderBar
 	/// <summary>
 	/// Represents the method that will handle a change in value.
 	/// </summary>
-	public delegate void ValueChangedHandler(object sender, decimal value);
+	public delegate void ValueChangedHandler(object sender, long value);
 
 	public enum MACBorderStyle
 	{
